requirements.txt
Flask
Flask-SQLAlchemy
Flask-Bcrypt
Flask-Session
PyJWT
cryptography
Werkzeug
pyotp
qrcode
pillow
python-magic-bin